Added sorting with case insensitive

+++ recon.sh
@@ -147,25 +147,35 @@ function permutation_subdomians(){
     
     axiom-scan $local/subdomain_results.txt -m dnsx-wc $domain -o $local/subdomain_perm_active.txt &> /dev/null
 
-    gotator -sub $local/subdomain_perm_active.txt -perm /opt/recon/lists/perm.txt -depth 1 -numbers 10 -mindup -adv -md -silent > $local/perm_gotator.txt
-      
-    axiom-scan $local/perm_gotator.txt -m dnsx-wc $domain -o $local/perm_resolved_sd.txt &> /dev/null
-    
-    cat $local/perm_resolved_sd.txt | anew $local/subdomain_results.txt > $local/perm_resolved_new.txt
-    
-    echo -n "${fgRed} --> Discovered $(wc -l <$local/perm_resolved_new.txt) new subdomains"
-    echo -n "${fgGreen} --> done"
-}
+    if [ ! $(wc -l < $local/recur1_active.txt) > 100 ]
+        then
+        gotator -sub $local/subdomain_perm_active.txt -perm /opt/recon/lists/perm.txt -depth 1 -numbers 10 -mindup -adv -md -silent > $local/perm_gotator.txt
+        
+        axiom-scan $local/perm_gotator.txt -m dnsx-wc $domain -o $local/perm_resolved_sd.txt &> /dev/null
+        
+        cat $local/perm_resolved_sd.txt | anew $local/subdomain_results.txt > $local/perm_resolved_new.txt
+        
+        echo -n "${fgRed} --> Discovered $(wc -l <$local/perm_resolved_new.txt) new subdomains"
+        echo -n "${fgGreen} --> done"
+
+    else
+        echo -n "${fgRed} --> Found more than 100 new subdomains - Skipping"
+        echo -n "${fgGreen} --> skipped"
+    fi
+
+    }
 
 function final_resolve() {
     
     axiom-select "scooby*" > /dev/null 2>&1
 
     echo -n "\n${fgYellow}Subdomain Enumeration : Resolving subdomains (final)${txReset} "
-    
-    axiom-scan $local/subdomain_results.txt -m dnsx -o $results_local/subdomain_active.txt &> /dev/null
 
-    axiom-scan $local/subdomain_active.txt -m dnsx -cname -resp -o $results_local/subdomain_cname.txt &> /dev/null
+    cat $local/subdomain_results.txt | sort -uf > $local/subdomain_final_sorted.txt
+    
+    axiom-scan $local/subdomain_final_sorted.txt -m dnsx -o $results_local/subdomain_active.txt &> /dev/null
+
+    axiom-scan $results_local/subdomain_active.txt -m dnsx -cname -resp -o $results_local/subdomain_cname.txt &> /dev/null
   
     echo -n "${fgRed} --> Discovered a total of $(wc -l <$results_local/subdomain_active.txt) new subdomains"
     echo -n "${fgGreen} --> done"
